+++ CHANGELOG.md
@@ -8,6 +8,9 @@ for the globale releases (`YYYY.MINOR`), and to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 
 ## [Unreleased]
+### Changed
+- `Text.font_files` text-based parameter is replaced with `Text.font_faces` node
+  list which accepts `FontFace` nodes instead
 
 ## [2024.0] [libnopegl 0.11.0] - 2024-02-02
 ### Added